Index/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-queue 
azure-search-documents
azure-cognitiveservices-vision-computervision
msrest
pillow
matplotlib
azure-core
